refactor: extract trace capability module (#363)
