+++ cmd/internal/serverapp/trace_composition.go
@@ -1,7 +1,7 @@
 package serverapp
 
-import "github.com/markhuangai/dense-mem/internal/service/contextservice"
+import "github.com/markhuangai/dense-mem/internal/trace"
 
-func buildContextApplication(store contextservice.SemanticTraceStore) contextservice.Service {
-	return contextservice.NewSemantic(store)
+func buildContextApplication(store trace.SemanticTraceStore) trace.Service {
+	return trace.NewSemantic(store)
 }
