+++ internal/trace/postgres/evidence_lifecycle_trace_repository.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"context"
