+++ internal/trace/postgres/semantic_read_repository_test.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"context"
