+++ internal/trace/postgres/semantic_read_search_repository.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"context"
