+++ internal/trace/postgres/semantic_trace_occurrence.go
@@ -1,4 +1,4 @@
-package repository
+package postgres
 
 import (
 	"context"
